MatrixAppMain.cpp (first upload)
This file contains the main file for the program matrix calculator and quiz program. It is still in its infant stage, but as of now, it supports quiz taking for matrix addition. Planned future features include matrix multiplication quizzes as well as transposition quizzes. The calculator is also planned for in future updates.