Update 10 (new functions and bug fix)
This update reflects the addition of two additional functions as well as a bug fix in one of other functions.
The bug fix is in the isValidRow function. It was not catching the case where the last character in a string was a minus sign, therefore making it invalid. This has been fixed.
One new function is named isEqual(Matrix &x) which returns true if two matrices are equal and false is otherwise.
The other new function is named getUserMatrix() which simply receives input from the user to store inside of a matrix. It utilizes the isValidRow function.