Update 1 (new menu functions & quiz option)
This update reflects the addition of new functions that make the code more readable and structured. It not contains the following four functions:

void displayMainMenu();
void displayMatrixOpsMenu();
void takeAdditionQuiz();
void takeMultiplicationQuiz();

The titles are descriptive of what they do. Their existence assists with better program flow, so it is easier to see what is going on when reading the code. Also, this update allows users to take the multiplication matrix test.